requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==3.1
packaging==20.9
pip-review==1.1.0
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
twilio==6.58.0
urllib3==1.26.4
